+++ fakeredis/_server.py
@@ -7,9 +7,9 @@ from typing import Dict, Tuple, Any, List, Optional, Union
 
 import redis
 
-from fakeredis.model import AccessControlList
 from fakeredis._helpers import Database, FakeSelector
-from fakeredis.typing import VersionType, ServerType
+from fakeredis._typing import VersionType, ServerType
+from fakeredis.model import AccessControlList
 
 LOGGER = logging.getLogger("fakeredis")
 
@@ -52,6 +52,7 @@ class FakeServer:
         self.lock = threading.Lock()
         self.dbs: Dict[int, Database] = defaultdict(lambda: Database(self.lock))
         # Maps channel/pattern to a weak set of sockets
+        self.script_cache: Dict[bytes, bytes] = dict()  # Maps SHA1 to the script source
         self.subscribers: Dict[bytes, weakref.WeakSet[Any]] = defaultdict(weakref.WeakSet)
         self.psubscribers: Dict[bytes, weakref.WeakSet[Any]] = defaultdict(weakref.WeakSet)
         self.ssubscribers: Dict[bytes, weakref.WeakSet[Any]] = defaultdict(weakref.WeakSet)
@@ -66,6 +67,14 @@ class FakeServer:
         self.server_type: str = server_type
         self.config: Dict[bytes, bytes] = config or dict()
         self.acl: AccessControlList = AccessControlList()
+        self.clients: Dict[str, Dict[str, Any]] = dict()
+        self._next_client_id = 1
+
+    def get_next_client_id(self) -> int:
+        with self.lock:
+            client_id = self._next_client_id
+            self._next_client_id += 1
+        return client_id
 
     @staticmethod
     def get_server(key: str, version: VersionType, server_type: ServerType) -> "FakeServer":